+++ CHANGELOG.md
@@ -77,6 +77,7 @@ With this release the systemd configuration files for InfluxDB will use the syst
 - [#6855](https://github.com/influxdata/influxdb/pull/6855): Update `stress/v2` to work with clusters, ssl, and username/password auth. Code cleanup
 - [#6738](https://github.com/influxdata/influxdb/issues/6738): Time sorting broken with overwritten points
 - [#6829](https://github.com/influxdata/influxdb/issues/6829): Fix panic: runtime error: index out of range
+- [#6911](https://github.com/influxdata/influxdb/issues/6911): Fix fill(previous) when used with math operators.
 
 ## v0.13.0 [2016-05-12]